fixing double to int and design issue

+++ lib/models/User.dart
@@ -13,7 +13,7 @@ class User {
     required this.token
   });
 
-  int budget;
+  double budget;
   String id;
   String email;
   String fullName;
